refactor(viewer): modify useCircleMeasurement directory location
