+++ libs/insight-viewer/src/Annotation/components/AreaDrawer/index.tsx
@@ -1,5 +1,5 @@
 import React, { ReactElement } from 'react'
-import useCircleMeasurement from '../../../hooks/useCircleMeasurement'
+import useCircleMeasurement from '../../hooks/useCircleMeasurement'
 
 import { svgWrapperStyle, textStyle } from '../../../Viewer/Viewer.styles'
