+++ libs/insight-viewer/src/Annotation/hooks/useCircleMeasurement/utils/getCircleTextPosition.spec.ts
@@ -1,4 +1,4 @@
-import { Point } from '../../../types'
+import { Point } from '../../../../types'
 import { getCircleTextPosition } from './getCircleTextPosition'
 
 describe('getCircleTextPosition: ', () => {
