+++ libs/insight-viewer/src/Annotation/utils/calculateCircleArea.spec.ts
