+++ libs/insight-viewer/src/Annotation/utils/calculateCircleArea.ts
